Added User Authentication in backend

+++ backend/controllers/goalController.js
@@ -2,12 +2,13 @@ const asyncHandler = require("express-async-handler"); // Had to install using n
 //  handler for us instead of using .then or try/catch. After this we can now use the mangoDB
 
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc    GET goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -24,6 +25,7 @@ const setGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.create({
     text: req.body.text,
     completed: false,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
@@ -43,6 +45,20 @@ const updateGoal = asyncHandler(async (req, res) => {
   // const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
   //   new: true,
   // });
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // make sure the logged in user matches the owner of the goal
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(
     req.params.id,
     { completed: !goal.completed },
@@ -64,6 +80,25 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // make sure the logged in user matches the owner of the goal
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
+  const updatedGoal = await Goal.findByIdAndUpdate(
+    req.params.id,
+    { completed: !goal.completed },
+    { new: true }
+  );
 
   await Goal.findByIdAndDelete(req.params.id);
